# Only use the token queue for fields we really can't parse immediately